pyproject.toml
[project]
name = "ol-pejeta-ground-control"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "pymavlink>=2.4.43",
    "pyserial>=3.5",
    "pyside6>=6.9.0",
    "python-multipart>=0.0.20",
    "ruff>=0.11.8",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]
